refactor: 매퍼 클래스 분리 및 위치 수정
