+++ app/src/main/java/woowacourse/shopping/data/util/Mapper.kt
@@ -1,0 +1,17 @@
+package woowacourse.shopping.data.util
+
+import woowacourse.shopping.data.local.CartEntity
+import woowacourse.shopping.domain.CartItem
+import woowacourse.shopping.domain.CartItems
+
+fun CartEntity.toDomain(): CartItem =
+    CartItem(
+        productId = productId,
+        amount = amount,
+    )
+
+fun List<CartEntity>.toDomain(isLast: Boolean): CartItems =
+    CartItems(
+        items = map { it.toDomain() },
+        isLast = isLast,
+    )
